backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
requests>=2.28.0
python-multipart>=0.0.6
python-pptx>=0.6.21
python-docx>=0.8.11
jinja2>=3.1.2
aiofiles>=23.1.0
tenacity>=8.0.0
reportlab>=3.6.12
aiohttp>=3.9.1 